requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.6.0
groq>=0.4.0
google-genai>=1.0.0
firebase-admin>=6.4.0
chromadb>=0.4.22
ddgs>=9.10.0
pymupdf4llm>=0.0.1
pymupdf>=1.23.0
rank_bm25>=0.2.2
requests>=2.31.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
python-multipart>=0.0.7
numpy<2.0.0
